Convert Drupal commands
